+++ .eslintrc.json
@@ -2,6 +2,7 @@
   "env": {
     "browser": false,
     "commonjs": true,
+    "node": true,
     "es6": true
   },
   "extends": "eslint:recommended",
