+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node index.js",
-    "lint": "eslint index.js"
+    "lint": "eslint ."
   },
   "repository": {
     "type": "git",
